ok looks like I need to avoid -Wold-style-cast too

+++ src/gui/widgets/text_box.hpp
@@ -104,7 +104,7 @@ public:
 
 private:
 	text_history(std::vector<std::string>* history, const bool enabled)
-		: history_(history), pos_(history->size()), enabled_(enabled)
+		: history_(history), pos_(static_cast<unsigned int>(history->size())), enabled_(enabled)
 	{
 	}
